Optionally disable threshold testing
As it happens, Home Assistant relies on the system-level bluetooth stack
to determine if a device has disappeared, meaning if
a BLE device goes away for more than 10 minutes it'll
mark all the attached sensors as Unavailable. Okay,
so let's stretch out the broadcast time and stop doing
range testing.